requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# plotting ------------------------------------

pandas>=1.1.4
seaborn>=0.11.0

# deep_sort -----------------------------------

easydict

# torchreid

Cython
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio


click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.22.1
opencv-python==4.5.3.56
six==1.16.0
Werkzeug==2.0.1
python-dotenv==0.19.2
redis==4.1.3


numpy==1.22.2
opencv-python==4.5.5.62

python-dotenv==0.19.2


youtube-dl==2021.12.17

redis==4.1.3